// src/app/shared/model/pages.model.ts
export interface pageSelection {
  skip: number;
  limit: number
}
// ...
export interface roles {
  name: string;
  shortName: string;
  uuid: string;
}

export interface SubModule {
  uuid: string;
  name: string;
  shortName: string;
  status: number;
  parentId: string;
}

export interface Module {
  uuid: string;
  name: string;
  shortName: string;
  submodules: SubModule[];
}

export interface userList {
  creditReference: number;
  userName: string;
  userId: string;
  email: string;
  mobile: string;
  status: number;
  balance: any;
  uuid: string;
}